More foramtting for small screen. Added initial command string.
Mostly re-formatted the self destruct sequence for the small screen. At least made a good start on it. Looks better now.

Put the initial command stuff back together and added it to arduinoTrekSetup() as a passed in strting. Works pretty slick. But, should it be used for subsequent games if the user stays in the game and keeps playing more games?