Create trek text.txt
Added the text version of the instructions.  I don't know what is a copy of what. Did the .doc file copy from the .txt version? Was the .txt version a copy  of the.doc file? I tend to think the .txt version has a greater chance of being the original.